move some functions in synthetic.py to func.py
