+++ obp/dataset/__init__.py
@@ -1,4 +1,5 @@
 from .base import *
+from .func import *
+from .multiclass import *
 from .real import *
 from .synthetic import *
-from .multiclass import *
